+++ LinearizedTrackFit/interface/TestFitter.h
@@ -18,7 +18,7 @@ namespace LinearFit
   void testFitter(const TString & inputFileName, const double & eventsFractionStart, const double & eventsFractionEnd,
       const std::vector<std::string> & inputVarNames, const std::vector<std::string> & inputTrackParameterNames,
       std::vector<double> & distanceCutsTransverse, std::vector<double> & distanceCutsLongitudinal,
-      std::unordered_map<int, std::pair<float, float> > & radiusCuts, bool singleModules, bool phiSymmetricFit)
+      std::unordered_map<int, std::pair<double, double> > & radiusCuts, bool singleModules, bool phiSymmetricFit)
   {
     std::vector<int> layersAll_{5, 6, 7, 8, 9, 10};
     std::unordered_map<std::string, std::unordered_set<int> > requiredLayers;
